+++ src/sphinx_gherkin_feature/__init__.py
@@ -10,7 +10,7 @@ from .domain import GherkinDomain
 from .lexer import GherkinFeatureLexer
 from .writer import write_feature_files
 
-__version__ = "0.1.1"
+__version__ = "0.1.3"
 
 
 def register_lexer(app: Sphinx) -> None: